requirements.txt
yfinance==0.2.40
prophet==1.1.5
feedparser==6.0.11
pystan==3.9.1
pandas==2.1.4
numpy==1.26.4
ta==0.11.0
firebase-admin==6.5.0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
pytz==2024.1
